[Refac] Carts query 분리하기
