+++ src/apis/user/index.ts
@@ -1,4 +1,4 @@
-import { cartListType, productListType } from '@/types/Product/productList';
+import { productListType } from '@/types/Product/productList';
 import axios from 'axios';
 
 export default async function getLikeList(): Promise<productListType> {
@@ -11,61 +11,6 @@ export default async function getLikeList(): Promise<productListType> {
     } catch (error) {
         console.error(error);
         throw Error;
-    }
-}
-export async function getCartList(): Promise<cartListType> {
-    try {
-        const response = await axios.get('/carts');
-        if (response.data) return response.data;
-        throw new Error(
-            `Unexpected response : ${response.status} ${response.statusText}`,
-        );
-    } catch (error) {
-        console.error(error);
-        throw Error;
-    }
-}
-
-export async function patchCart(data: { checked: boolean }) {
-    try {
-        const response = await axios.patch(`/carts`, data);
-        if (response.data) return response.data;
-        throw new Error(
-            `Unexpected response: ${response.status} ${response.statusText}`,
-        );
-    } catch (error) {
-        if (axios.isAxiosError(error)) {
-            console.error(
-                'Axios error:',
-                error.response?.data || error.message,
-            );
-            throw new Error(error.response?.data?.message || error.message);
-        }
-        console.error('Unknown error:', error);
-        throw new Error('An unknown error occurred');
-    }
-}
-
-export async function patchCartItem(data: {
-    productSeq: number;
-    checked: boolean;
-}) {
-    try {
-        const response = await axios.patch(`/carts/${data.productSeq}`, data);
-        if (response.data) return response.data;
-        throw new Error(
-            `Unexpected response: ${response.status} ${response.statusText}`,
-        );
-    } catch (error) {
-        if (axios.isAxiosError(error)) {
-            console.error(
-                'Axios error:',
-                error.response?.data || error.message,
-            );
-            throw new Error(error.response?.data?.message || error.message);
-        }
-        console.error('Unknown error:', error);
-        throw new Error('An unknown error occurred');
     }
 }
