+++ src/hooks/query/carts/index.ts
@@ -1,0 +1,15 @@
+import { postInCart } from '@/apis/carts';
+import { useMutation } from '@tanstack/react-query';
+
+export function usePostInCart() {
+    const { mutate: mutatePostInCart, isPending: isPendingPostInCart } =
+        useMutation({
+            mutationFn: (data: { productSeq: number; quantity: number }) =>
+                postInCart(data),
+        });
+
+    return {
+        mutatePostInCart,
+        isPendingPostInCart,
+    };
+}
